+++ examples/component/dropdown.cpp
@@ -4,9 +4,13 @@
 #include <string>  // for basic_string, string, allocator
 #include <vector>  // for vector
 
+#ifndef FTXUI_BUILD_MODULES
 #include "ftxui/component/captured_mouse.hpp"  // for ftxui
 #include "ftxui/component/component.hpp"  // for Dropdown, Horizontal, Vertical
 #include "ftxui/component/screen_interactive.hpp"  // for ScreenInteractive
+#else
+import ftxui.component;
+#endif
 
 int main() {
   using namespace ftxui;