+++ examples/dom/separator_style.cpp
@@ -1,13 +1,21 @@
 // Copyright 2020 Arthur Sonzogni. All rights reserved.
 // Use of this source code is governed by the MIT license that can be found in
 // the LICENSE file.
+#ifndef FTXUI_BUILD_MODULES
 #include <ftxui/screen/screen.hpp>  // for Screen
+#endif 
+
 #include <iostream>                 // for endl, cout, ostream
 #include <memory>                   // for allocator
 
+#ifndef FTXUI_BUILD_MODULES
 #include "ftxui/dom/elements.hpp"  // for text, hbox, separatorDouble, separatorHeavy, separatorLight, vbox, operator|, Element, Fit, borderDouble, borderHeavy, borderLight
 #include "ftxui/dom/node.hpp"      // for Render
 #include "ftxui/screen/box.hpp"    // for ftxui
+#else
+import ftxui.dom;
+import ftxui.screen;
+#endif
 
 int main() {
   using namespace ftxui;